+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # Change Log - oav
 
+## 01/15/2023 3.3.1
+
+- Add json report parameter in command `validate-traffic`
+
 ## 12/19/2022 3.3.0
 
 - API Scenario
@@ -11,6 +15,7 @@
 
 ## 12/01/2022 3.2.4
 
+- Generate Examples - Support specified generation rule
 - API Scenario
   - Support file type in formdata and body
   - Change the prefix length to 8 from 10 in generated API Scenario
@@ -490,7 +495,7 @@
 
 ## 02/13/2020 0.20.9
 
-- Fall back to match child resouce in live validator
+- Fall back to match child resource in live validator
 
 ## 02/13/2020 0.20.8